Make the `Dictionary` methods return token count
The token count of the matching entry in the dictionary allows to
optimize not-yet-implemented feature vector evaluation by skiping tokens
that match the tails of the found entry. I am assuming that if some
series of tokens match an entry in the dictionary then none of the
subseries matches any other entry.